requirements.txt
torch==2.5.1
ultralytics==8.3.35
easyocr==1.7.2
opencv-python==4.10.0.82
pandas==2.2.1
xgboost==2.1.2
lightgbm==4.5.0
pyyaml==6.0.1
scikit-learn==1.5.2
flask==3.0.2
gunicorn==23.0.0